Move phone counts computation up to increase render speed
