+++ src/components/StudentList/StudentList.tsx
@@ -1,6 +1,8 @@
-import React, { RefObject } from "react";
+import React, { RefObject, useMemo } from "react";
 import { StudentCard } from "..";
+import { useStudentStore } from "../../hooks";
 import { Student } from "../../interfaces";
+import { getActivePrimaryPhoneCounts, getPhoneCounts } from "../../services";
 import { VirtualizedList } from "../reusables";
 
 interface StudentListProps {
@@ -14,9 +16,23 @@ export const StudentList: React.FC<StudentListProps> = ({
   handleStudentDialogOpen,
   menuRef,
 }) => {
+  const students = useStudentStore((state) => {
+    return state.students;
+  });
+  const phoneCounts = useMemo(() => {
+    return getPhoneCounts(students);
+  }, [students]);
+  const primaryPhoneCounts = useMemo(() => {
+    return getActivePrimaryPhoneCounts(students);
+  }, [students]);
+
   return (
     <VirtualizedList idPath="epId" listData={filteredStudents} menuRef={menuRef} overscan={500}>
-      <StudentCard handleStudentDialogOpen={handleStudentDialogOpen} />
+      <StudentCard
+        handleStudentDialogOpen={handleStudentDialogOpen}
+        phoneCounts={phoneCounts}
+        primaryPhoneCounts={primaryPhoneCounts}
+      />
     </VirtualizedList>
   );
 };
